Refactor WizClient and models into partial classes
