+++ WizCloud.Tests/TokenRefreshTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace WizCloud.Tests;
@@ -9,8 +10,8 @@ public sealed class TokenRefreshTests {
     [TestMethod]
     public void WizClient_RefreshesTokenOnUnauthorized() {
         var repoRoot = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
-        var filePath = Path.Combine(repoRoot, "WizCloud", "WizClient.cs");
-        var source = File.ReadAllText(filePath);
+        var directory = Path.Combine(repoRoot, "WizCloud");
+        var source = string.Concat(Directory.GetFiles(directory, "WizClient*.cs").Select(File.ReadAllText));
         StringAssert.Contains(source, "AcquireTokenAsync");
         StringAssert.Contains(source, "Unauthorized");
     }
@@ -18,8 +19,8 @@ public sealed class TokenRefreshTests {
     [TestMethod]
     public void GetCloudAccounts_UsesSendWithRefreshAsync() {
         var repoRoot = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
-        var filePath = Path.Combine(repoRoot, "WizCloud", "WizClient.cs");
-        var source = File.ReadAllText(filePath);
+        var directory = Path.Combine(repoRoot, "WizCloud");
+        var source = string.Concat(Directory.GetFiles(directory, "WizClient*.cs").Select(File.ReadAllText));
         var index = source.IndexOf("GetCloudAccountsPageAsync", StringComparison.Ordinal);
         Assert.IsTrue(index >= 0, "GetCloudAccountsPageAsync method not found");
         var callIndex = source.IndexOf("SendWithRefreshAsync", index, StringComparison.Ordinal);
