+++ WizCloud.Tests/WizClientErrorHandlingTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.IO;
+using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace WizCloud.Tests;
@@ -9,8 +10,8 @@ public sealed class WizClientErrorHandlingTests {
     [TestMethod]
     public void WizClient_ContainsDetailedErrorMessage() {
         var repoRoot = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
-        var filePath = Path.Combine(repoRoot, "WizCloud", "WizClient.cs");
-        var source = File.ReadAllText(filePath);
+        var directory = Path.Combine(repoRoot, "WizCloud");
+        var source = string.Concat(Directory.GetFiles(directory, "WizClient*.cs").Select(File.ReadAllText));
         StringAssert.Contains(source, "Request failed with status code");
     }
 }
