+++ WizCloud.Tests/WizClientGraphQlQueriesTests.cs
@@ -1,6 +1,7 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.IO;
+using System.Linq;
 
 namespace WizCloud.Tests;
 
@@ -9,8 +10,8 @@ public sealed class WizClientGraphQlQueriesTests {
     [TestMethod]
     public void WizClient_UsesGraphQlQueryConstants() {
         var repoRoot = Path.GetFullPath(Path.Combine(AppContext.BaseDirectory, "..", "..", "..", ".."));
-        var filePath = Path.Combine(repoRoot, "WizCloud", "WizClient.cs");
-        var source = File.ReadAllText(filePath);
+        var directory = Path.Combine(repoRoot, "WizCloud");
+        var source = string.Concat(Directory.GetFiles(directory, "WizClient*.cs").Select(File.ReadAllText));
         StringAssert.Contains(source, "GraphQlQueries.UsersQuery");
        StringAssert.Contains(source, "GraphQlQueries.ProjectsQuery");
        StringAssert.Contains(source, "GraphQlQueries.CloudAccountsQuery");
