+++ WizCloud/WizClient.cs
@@ -16,7 +16,7 @@ namespace WizCloud;
 /// <summary>
 /// Provides a client for interacting with the Wiz GraphQL API.
 /// </summary>
-public class WizClient : IDisposable {
+public partial class WizClient : IDisposable {
     private static readonly HttpClient _httpClient = CreateClient();
     private readonly string _apiEndpoint;
     private readonly string? _clientId;
@@ -88,777 +88,6 @@ public class WizClient : IDisposable {
 
         return response;
     }
-
-    /// <summary>
-    /// Retrieves all users from Wiz asynchronously.
-    /// </summary>
-    /// <param name="pageSize">The number of users to retrieve per page. Defaults to 20.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains a list of all users.</returns>
-    public async Task<List<WizUser>> GetUsersAsync(
-        int pageSize = 20,
-        IEnumerable<WizUserType>? types = null,
-        string? projectId = null) {
-        var users = new List<WizUser>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetUsersPageAsync(pageSize, endCursor, types, projectId);
-            users.AddRange(result.Users);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return users;
-    }
-
-    /// <summary>
-    /// Streams users from Wiz asynchronously as an <see cref="IAsyncEnumerable{WizUser}"/>.
-    /// </summary>
-    /// <param name="pageSize">The number of users to retrieve per page. Defaults to 20.</param>
-    /// <param name="cancellationToken">Token to monitor for cancellation requests.</param>
-    /// <returns>An async enumerable sequence of users.</returns>
-    public async IAsyncEnumerable<WizUser> GetUsersAsyncEnumerable(
-        int pageSize = 20,
-        IEnumerable<WizUserType>? types = null,
-        string? projectId = null,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            (List<WizUser> Users, bool HasNextPage, string? EndCursor) result;
-            try {
-                result = await GetUsersPageAsync(pageSize, endCursor, types, projectId).ConfigureAwait(false);
-            } catch (HttpRequestException) {
-                yield break;
-            }
-
-            foreach (var user in result.Users) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return user;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of users from the Wiz API.
-    /// </summary>
-    /// <param name="first">The number of users to retrieve.</param>
-    /// <param name="after">The cursor for pagination, if retrieving subsequent pages.</param>
-    /// <returns>A tuple containing the users, whether there's a next page, and the cursor for the next page.</returns>
-    private async Task<(List<WizUser> Users, bool HasNextPage, string? EndCursor)> GetUsersPageAsync(
-        int first,
-        string? after = null,
-        IEnumerable<WizUserType>? types = null,
-        string? projectId = null) {
-        var query = GraphQlQueries.UsersQuery;
-
-        var typeFilter = types != null && types.Any()
-            ? types.Select(t => t.ToString())
-            : new[] { "USER_ACCOUNT", "SERVICE_ACCOUNT", "GROUP", "ACCESS_KEY" };
-
-        var propertyFilter = projectId != null
-            ? new object[] { new { name = "projectId", equals = new[] { projectId } } }
-            : Array.Empty<object>();
-
-        var variables = new {
-            first,
-            after,
-            filterBy = new {
-                type = new { equals = typeFilter },
-                property = propertyFilter
-            }
-        };
-
-        var requestBody = new {
-            query,
-            variables
-        };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                if (!response.IsSuccessStatusCode) {
-                    var errorBody = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    var message = $"Request failed with status code {(int)response.StatusCode} ({response.ReasonPhrase}).";
-                    if (!string.IsNullOrWhiteSpace(errorBody))
-                        message += $" Body: {errorBody}";
-                    throw new HttpRequestException(message);
-                }
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var users = new List<WizUser>();
-                var nodes = jsonResponse["data"]?["cloudResourcesV2"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null)
-                            users.Add(WizUser.FromJson(node));
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["cloudResourcesV2"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (users ?? new List<WizUser>(), hasNextPage, endCursor);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all projects from Wiz asynchronously.
-    /// </summary>
-    /// <param name="pageSize">The number of projects to retrieve per page. Defaults to 20.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains a list of all projects.</returns>
-    public async Task<List<WizProject>> GetProjectsAsync(int pageSize = 20) {
-        var projects = new List<WizProject>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetProjectsPageAsync(pageSize, endCursor);
-            projects.AddRange(result.Projects);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return projects;
-    }
-
-    /// <summary>
-    /// Streams projects from Wiz asynchronously as an <see cref="IAsyncEnumerable{WizProject}"/>.
-    /// </summary>
-    /// <param name="pageSize">The number of projects to retrieve per page. Defaults to 20.</param>
-    /// <param name="cancellationToken">Token to monitor for cancellation requests.</param>
-    /// <returns>An async enumerable sequence of projects.</returns>
-    public async IAsyncEnumerable<WizProject> GetProjectsAsyncEnumerable(int pageSize = 20, [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            var result = await GetProjectsPageAsync(pageSize, endCursor).ConfigureAwait(false);
-
-            foreach (var project in result.Projects) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return project;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of projects from the Wiz API.
-    /// </summary>
-    /// <param name="first">The number of projects to retrieve.</param>
-    /// <param name="after">The cursor for pagination, if retrieving subsequent pages.</param>
-    /// <returns>A tuple containing the projects, whether there's a next page, and the cursor for the next page.</returns>
-    private async Task<(List<WizProject> Projects, bool HasNextPage, string? EndCursor)> GetProjectsPageAsync(int first, string? after = null) {
-        const string query = GraphQlQueries.ProjectsQuery;
-
-        var variables = new {
-            first,
-            after
-        };
-
-        var requestBody = new {
-            query,
-            variables
-        };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                response.EnsureSuccessStatusCode();
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var projects = new List<WizProject>();
-                var nodes = jsonResponse["data"]?["projects"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null) {
-                            projects.Add(new WizProject {
-                                Id = node["id"]?.GetValue<string>() ?? string.Empty,
-                                Name = node["name"]?.GetValue<string>() ?? string.Empty,
-                                Slug = node["slug"]?.GetValue<string>() ?? string.Empty,
-                                IsFolder = node["isFolder"]?.GetValue<bool>() ?? false
-                            });
-                        }
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["projects"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (projects, hasNextPage, endCursor);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all cloud accounts from Wiz asynchronously.
-    /// </summary>
-    /// <param name="pageSize">The number of cloud accounts to retrieve per page. Defaults to 20.</param>
-    /// <returns>A task that represents the asynchronous operation. The task result contains a list of all cloud accounts.</returns>
-    public async Task<List<WizCloudAccount>> GetCloudAccountsAsync(int pageSize = 20) {
-        var accounts = new List<WizCloudAccount>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetCloudAccountsPageAsync(pageSize, endCursor).ConfigureAwait(false);
-            accounts.AddRange(result.Accounts);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return accounts;
-    }
-
-    /// <summary>
-    /// Streams cloud accounts from Wiz asynchronously as an <see cref="IAsyncEnumerable{WizCloudAccount}"/>.
-    /// </summary>
-    /// <param name="pageSize">The number of cloud accounts to retrieve per page. Defaults to 20.</param>
-    /// <param name="cancellationToken">Token to monitor for cancellation requests.</param>
-    /// <returns>An async enumerable sequence of cloud accounts.</returns>
-    public async IAsyncEnumerable<WizCloudAccount> GetCloudAccountsAsyncEnumerable(int pageSize = 20, [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            var result = await GetCloudAccountsPageAsync(pageSize, endCursor).ConfigureAwait(false);
-
-            foreach (var account in result.Accounts) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return account;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of cloud accounts from the Wiz API.
-    /// </summary>
-    /// <param name="first">The number of cloud accounts to retrieve.</param>
-    /// <param name="after">The cursor for pagination, if retrieving subsequent pages.</param>
-    /// <returns>A tuple containing the cloud accounts, whether there's a next page, and the cursor for the next page.</returns>
-    private async Task<(List<WizCloudAccount> Accounts, bool HasNextPage, string? EndCursor)> GetCloudAccountsPageAsync(int first, string? after = null) {
-        const string query = GraphQlQueries.CloudAccountsQuery;
-
-        var variables = new {
-            first,
-            after
-        };
-
-        var requestBody = new {
-            query,
-            variables
-        };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _token);
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                response.EnsureSuccessStatusCode();
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var accounts = new List<WizCloudAccount>();
-                var nodes = jsonResponse["data"]?["cloudAccounts"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null) {
-                            accounts.Add(WizCloudAccount.FromJson(node));
-                        }
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["cloudAccounts"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (accounts, hasNextPage, endCursor);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all issues from Wiz asynchronously.
-    /// </summary>
-    public async Task<List<WizIssue>> GetIssuesAsync(
-        int pageSize = 20,
-        IEnumerable<WizSeverity>? severities = null,
-        IEnumerable<string>? statuses = null,
-        string? projectId = null,
-        IEnumerable<string>? types = null) {
-        var issues = new List<WizIssue>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetIssuesPageAsync(pageSize, endCursor, severities, statuses, projectId, types).ConfigureAwait(false);
-            issues.AddRange(result.Issues);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return issues;
-    }
-
-    /// <summary>
-    /// Streams issues from Wiz asynchronously.
-    /// </summary>
-    public async IAsyncEnumerable<WizIssue> GetIssuesAsyncEnumerable(
-        int pageSize = 20,
-        IEnumerable<WizSeverity>? severities = null,
-        IEnumerable<string>? statuses = null,
-        string? projectId = null,
-        IEnumerable<string>? types = null,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            (List<WizIssue> Issues, bool HasNextPage, string? EndCursor) result;
-            try {
-                result = await GetIssuesPageAsync(pageSize, endCursor, severities, statuses, projectId, types).ConfigureAwait(false);
-            } catch (HttpRequestException) {
-                yield break;
-            }
-
-            foreach (var issue in result.Issues) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return issue;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of issues from the Wiz API.
-    /// </summary>
-    private async Task<(List<WizIssue> Issues, bool HasNextPage, string? EndCursor)> GetIssuesPageAsync(
-        int first,
-        string? after = null,
-        IEnumerable<WizSeverity>? severities = null,
-        IEnumerable<string>? statuses = null,
-        string? projectId = null,
-        IEnumerable<string>? types = null) {
-        const string query = GraphQlQueries.IssuesQuery;
-
-        var severityFilter = severities != null && severities.Any()
-            ? new { equals = severities.Select(s => s.ToString()) }
-            : null;
-        var statusFilter = statuses != null && statuses.Any()
-            ? new { equals = statuses }
-            : null;
-        var typeFilter = types != null && types.Any()
-            ? new { equals = types }
-            : null;
-        var projectFilter = projectId != null ? new { equals = new[] { projectId } } : null;
-
-        var variables = new {
-            first,
-            after,
-            filterBy = new {
-                severity = severityFilter,
-                status = statusFilter,
-                type = typeFilter,
-                projectId = projectFilter
-            }
-        };
-
-        var requestBody = new { query, variables };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _token);
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                if (!response.IsSuccessStatusCode) {
-                    var errorBody = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    var message = $"Request failed with status code {(int)response.StatusCode} ({response.ReasonPhrase}).";
-                    if (!string.IsNullOrWhiteSpace(errorBody))
-                        message += $" Body: {errorBody}";
-                    throw new HttpRequestException(message);
-                }
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var issues = new List<WizIssue>();
-                var nodes = jsonResponse["data"]?["issues"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null)
-                            issues.Add(WizIssue.FromJson(node));
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["issues"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (issues, hasNextPage, endCursor);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all vulnerabilities from Wiz asynchronously.
-    /// </summary>
-    public async Task<List<WizVulnerability>> GetVulnerabilitiesAsync(
-        int pageSize = 20,
-        string? cve = null,
-        double? minCvss = null,
-        bool? exploitAvailable = null,
-        string? projectId = null) {
-        var vulnerabilities = new List<WizVulnerability>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetVulnerabilitiesPageAsync(
-                pageSize,
-                endCursor,
-                cve,
-                minCvss,
-                exploitAvailable,
-                projectId).ConfigureAwait(false);
-            vulnerabilities.AddRange(result.Vulnerabilities);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return vulnerabilities;
-    }
-
-    /// <summary>
-    /// Streams vulnerabilities from Wiz asynchronously.
-    /// </summary>
-    public async IAsyncEnumerable<WizVulnerability> GetVulnerabilitiesAsyncEnumerable(
-        int pageSize = 20,
-        string? cve = null,
-        double? minCvss = null,
-        bool? exploitAvailable = null,
-        string? projectId = null,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            (List<WizVulnerability> Vulnerabilities, bool HasNextPage, string? EndCursor) result;
-            try {
-                result = await GetVulnerabilitiesPageAsync(
-                    pageSize,
-                    endCursor,
-                    cve,
-                    minCvss,
-                    exploitAvailable,
-                    projectId).ConfigureAwait(false);
-            } catch (HttpRequestException) {
-                yield break;
-            }
-
-            foreach (var vulnerability in result.Vulnerabilities) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return vulnerability;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of vulnerabilities from the Wiz API.
-    /// </summary>
-    private async Task<(List<WizVulnerability> Vulnerabilities, bool HasNextPage, string? EndCursor)> GetVulnerabilitiesPageAsync(
-        int first,
-        string? after = null,
-        string? cve = null,
-        double? minCvss = null,
-        bool? exploitAvailable = null,
-        string? projectId = null) {
-        const string query = GraphQlQueries.VulnerabilitiesQuery;
-
-        var cveFilter = cve != null ? new { equals = new[] { cve } } : null;
-        var cvssFilter = minCvss != null ? new { score = new { gte = minCvss } } : null;
-        var exploitFilter = exploitAvailable.HasValue ? new { equals = exploitAvailable.Value } : null;
-        var projectFilter = projectId != null ? new { equals = new[] { projectId } } : null;
-
-        var variables = new {
-            first,
-            after,
-            filterBy = new {
-                cve = cveFilter,
-                cvss = cvssFilter,
-                exploitAvailable = exploitFilter,
-                projectId = projectFilter
-            }
-        };
-
-        var requestBody = new { query, variables };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                if (!response.IsSuccessStatusCode) {
-                    var errorBody = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    var message = $"Request failed with status code {(int)response.StatusCode} ({response.ReasonPhrase}).";
-                    if (!string.IsNullOrWhiteSpace(errorBody))
-                        message += $" Body: {errorBody}";
-                    throw new HttpRequestException(message);
-                }
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var vulnerabilities = new List<WizVulnerability>();
-                var nodes = jsonResponse["data"]?["vulnerabilities"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null)
-                            vulnerabilities.Add(WizVulnerability.FromJson(node));
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["vulnerabilities"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (vulnerabilities, hasNextPage, endCursor);
-            }
-        }
-    }
-
-    /// <summary>
-    /// Retrieves all resources from Wiz asynchronously.
-    /// </summary>
-    public async Task<List<WizResource>> GetResourcesAsync(
-        int pageSize = 20,
-        IEnumerable<string>? types = null,
-        IEnumerable<WizCloudProvider>? cloudProviders = null,
-        string? region = null,
-        bool? publiclyAccessible = null,
-        IDictionary<string, string>? tags = null,
-        string? projectId = null) {
-        var resources = new List<WizResource>();
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (hasNextPage) {
-            var result = await GetResourcesPageAsync(
-                pageSize,
-                endCursor,
-                types,
-                cloudProviders,
-                region,
-                publiclyAccessible,
-                tags,
-                projectId).ConfigureAwait(false);
-            resources.AddRange(result.Resources);
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-
-        return resources;
-    }
-
-    /// <summary>
-    /// Streams resources from Wiz asynchronously.
-    /// </summary>
-    public async IAsyncEnumerable<WizResource> GetResourcesAsyncEnumerable(
-        int pageSize = 20,
-        IEnumerable<string>? types = null,
-        IEnumerable<WizCloudProvider>? cloudProviders = null,
-        string? region = null,
-        bool? publiclyAccessible = null,
-        IDictionary<string, string>? tags = null,
-        string? projectId = null,
-        [EnumeratorCancellation] CancellationToken cancellationToken = default) {
-        string? endCursor = null;
-        bool hasNextPage = true;
-
-        while (!cancellationToken.IsCancellationRequested && hasNextPage) {
-            (List<WizResource> Resources, bool HasNextPage, string? EndCursor) result;
-            try {
-                result = await GetResourcesPageAsync(
-                    pageSize,
-                    endCursor,
-                    types,
-                    cloudProviders,
-                    region,
-                    publiclyAccessible,
-                    tags,
-                    projectId).ConfigureAwait(false);
-            } catch (HttpRequestException) {
-                yield break;
-            }
-
-            foreach (var resource in result.Resources) {
-                if (cancellationToken.IsCancellationRequested)
-                    yield break;
-
-                yield return resource;
-            }
-
-            hasNextPage = result.HasNextPage;
-            endCursor = result.EndCursor;
-        }
-    }
-
-    /// <summary>
-    /// Retrieves a single page of resources from the Wiz API.
-    /// </summary>
-    private async Task<(List<WizResource> Resources, bool HasNextPage, string? EndCursor)> GetResourcesPageAsync(
-        int first,
-        string? after = null,
-        IEnumerable<string>? types = null,
-        IEnumerable<WizCloudProvider>? cloudProviders = null,
-        string? region = null,
-        bool? publiclyAccessible = null,
-        IDictionary<string, string>? tags = null,
-        string? projectId = null) {
-        const string query = GraphQlQueries.ResourcesQuery;
-
-        var typeFilter = types != null && types.Any() ? new { equals = types } : null;
-        var cloudProviderFilter = cloudProviders != null && cloudProviders.Any()
-            ? new { equals = cloudProviders.Select(cp => cp.ToString()) }
-            : null;
-        var regionFilter = region != null ? new { equals = new[] { region } } : null;
-        var publicFilter = publiclyAccessible.HasValue ? new { equals = publiclyAccessible.Value } : null;
-        var tagFilter = tags != null && tags.Count > 0
-            ? tags.Select(kvp => new { name = kvp.Key, equals = new[] { kvp.Value } }).ToArray()
-            : Array.Empty<object>();
-        var projectFilter = projectId != null ? new { equals = new[] { projectId } } : null;
-
-        var variables = new {
-            first,
-            after,
-            filterBy = new {
-                type = typeFilter,
-                cloudPlatform = cloudProviderFilter,
-                region = regionFilter,
-                publiclyAccessible = publicFilter,
-                tags = tagFilter,
-                projectId = projectFilter
-            }
-        };
-
-        var requestBody = new { query, variables };
-
-        using (var request = new HttpRequestMessage(HttpMethod.Post, _apiEndpoint)) {
-            request.Headers.Authorization = new AuthenticationHeaderValue("Bearer", _token);
-            request.Content = new StringContent(
-                JsonSerializer.Serialize(requestBody),
-                Encoding.UTF8,
-                "application/json"
-            );
-
-            using (var response = await SendWithRefreshAsync(request).ConfigureAwait(false)) {
-                if (!response.IsSuccessStatusCode) {
-                    var errorBody = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                    var message = $"Request failed with status code {(int)response.StatusCode} ({response.ReasonPhrase}).";
-                    if (!string.IsNullOrWhiteSpace(errorBody))
-                        message += $" Body: {errorBody}";
-                    throw new HttpRequestException(message);
-                }
-
-                var content = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
-                var jsonResponse = JsonNode.Parse(content);
-
-                if (jsonResponse == null)
-                    throw new InvalidOperationException("Received null response from API");
-
-                var resources = new List<WizResource>();
-                var nodes = jsonResponse["data"]?["resources"]?["nodes"]?.AsArray();
-
-                if (nodes != null) {
-                    foreach (var node in nodes) {
-                        if (node != null)
-                            resources.Add(WizResource.FromJson(node));
-                    }
-                }
-
-                var pageInfo = jsonResponse["data"]?["resources"]?["pageInfo"];
-                bool hasNextPage = pageInfo?["hasNextPage"]?.GetValue<bool>() ?? false;
-                string? endCursor = pageInfo?["endCursor"]?.GetValue<string>();
-
-                return (resources, hasNextPage, endCursor);
-            }
-        }
-    }
-
     /// <summary>
     /// Releases all resources used by the WizClient.
     /// </summary>
